fix(frontend): review comments
- use selector component loading props
- move fethcer logic to `fetchers` folder
